python_ai/requirements.txt
flask
flask_sock
llama-cpp-python>=0.2.43  # optional, CPU build by default 